# perf: zero-copy tokenizer