--- a/go/vt/binlog/event_streamer.go
+++ b/go/vt/binlog/event_streamer.go
@@ -200,7 +200,7 @@ func parsePkNames(tokenizer *sqlparser.Tokenizer) ([]*querypb.Field, error) {
 				Name: string(val),
 			})
 		default:
-			return nil, fmt.Errorf("syntax error at position: %d", tokenizer.Position)
+			return nil, fmt.Errorf("syntax error at position: %d", tokenizer.Pos)
 		}
 	}
 	return columns, nil
@@ -297,15 +297,14 @@ func parsePkTuple(tokenizer *sqlparser.Tokenizer, insertid int64, fields []*quer
 				return nil, insertid, fmt.Errorf("incompatible string field with type %v", fields[index].Type)
 			}
 
-			decoded := make([]byte, base64.StdEncoding.DecodedLen(len(val)))
-			numDecoded, err := base64.StdEncoding.Decode(decoded, val)
+			decoded, err := base64.StdEncoding.DecodeString(val)
 			if err != nil {
 				return nil, insertid, err
 			}
-			result.Lengths = append(result.Lengths, int64(numDecoded))
-			result.Values = append(result.Values, decoded[:numDecoded]...)
+			result.Lengths = append(result.Lengths, int64(len(decoded)))
+			result.Values = append(result.Values, decoded...)
 		default:
-			return nil, insertid, fmt.Errorf("syntax error at position: %d", tokenizer.Position)
+			return nil, insertid, fmt.Errorf("syntax error at position: %d", tokenizer.Pos)
 		}
 		index++
 	}