--- a/go/vt/vtgate/planbuilder/expr.go
+++ b/go/vt/vtgate/planbuilder/expr.go
@@ -156,7 +156,7 @@ func (pb *primitiveBuilder) findOrigin(expr sqlparser.Expr) (pullouts []*pullout
 		construct, ok := constructsMap[sqi.ast]
 		if !ok {
 			// (subquery) -> :_sq
-			expr = sqlparser.ReplaceExpr(expr, sqi.ast, sqlparser.NewArgument([]byte(":"+sqName)))
+			expr = sqlparser.ReplaceExpr(expr, sqi.ast, sqlparser.NewArgument(":"+sqName))
 			pullouts = append(pullouts, newPulloutSubquery(engine.PulloutValue, sqName, hasValues, sqi.plan))
 			continue
 		}
@@ -170,9 +170,9 @@ func (pb *primitiveBuilder) findOrigin(expr sqlparser.Expr) (pullouts []*pullout
 					Right:    sqlparser.ListArg("::" + sqName),
 				}
 				left := &sqlparser.ComparisonExpr{
-					Left:     sqlparser.NewArgument([]byte(":" + hasValues)),
+					Left:     sqlparser.NewArgument(":" + hasValues),
 					Operator: sqlparser.EqualOp,
-					Right:    sqlparser.NewIntLiteral([]byte("1")),
+					Right:    sqlparser.NewIntLiteral("1"),
 				}
 				newExpr := &sqlparser.AndExpr{
 					Left:  left,
@@ -183,9 +183,9 @@ func (pb *primitiveBuilder) findOrigin(expr sqlparser.Expr) (pullouts []*pullout
 			} else {
 				// a not in (subquery) -> (:__sq_has_values = 0 or (a not in ::__sq))
 				left := &sqlparser.ComparisonExpr{
-					Left:     sqlparser.NewArgument([]byte(":" + hasValues)),
+					Left:     sqlparser.NewArgument(":" + hasValues),
 					Operator: sqlparser.EqualOp,
-					Right:    sqlparser.NewIntLiteral([]byte("0")),
+					Right:    sqlparser.NewIntLiteral("0"),
 				}
 				right := &sqlparser.ComparisonExpr{
 					Operator: construct.Operator,
@@ -201,7 +201,7 @@ func (pb *primitiveBuilder) findOrigin(expr sqlparser.Expr) (pullouts []*pullout
 			}
 		case *sqlparser.ExistsExpr:
 			// exists (subquery) -> :__sq_has_values
-			expr = sqlparser.ReplaceExpr(expr, construct, sqlparser.NewArgument([]byte(":"+hasValues)))
+			expr = sqlparser.ReplaceExpr(expr, construct, sqlparser.NewArgument(":"+hasValues))
 			pullouts = append(pullouts, newPulloutSubquery(engine.PulloutExists, sqName, hasValues, sqi.plan))
 		}
 	}
@@ -299,7 +299,7 @@ func valEqual(a, b sqlparser.Expr) bool {
 		if !ok {
 			return false
 		}
-		return bytes.Equal(a, b)
+		return a == b
 	case *sqlparser.Literal:
 		b, ok := b.(*sqlparser.Literal)
 		if !ok {
@@ -309,7 +309,7 @@ func valEqual(a, b sqlparser.Expr) bool {
 		case sqlparser.StrVal:
 			switch b.Type {
 			case sqlparser.StrVal:
-				return bytes.Equal(a.Val, b.Val)
+				return a.Val == b.Val
 			case sqlparser.HexVal:
 				return hexEqual(b, a)
 			}
@@ -317,7 +317,7 @@ func valEqual(a, b sqlparser.Expr) bool {
 			return hexEqual(a, b)
 		case sqlparser.IntVal:
 			if b.Type == (sqlparser.IntVal) {
-				return bytes.Equal(a.Val, b.Val)
+				return a.Val == b.Val
 			}
 		}
 	}
@@ -331,7 +331,7 @@ func hexEqual(a, b *sqlparser.Literal) bool {
 	}
 	switch b.Type {
 	case sqlparser.StrVal:
-		return bytes.Equal(v, b.Val)
+		return bytes.Equal(v, b.Bytes())
 	case sqlparser.HexVal:
 		v2, err := b.HexDecode()
 		if err != nil {