--- a/go/vt/vtgate/planbuilder/expression_converter.go
+++ b/go/vt/vtgate/planbuilder/expression_converter.go
@@ -38,7 +38,7 @@ func booleanValues(astExpr sqlparser.Expr) evalengine.Expr {
 	case *sqlparser.Literal:
 		//set autocommit = 'on'
 		if node.Type == sqlparser.StrVal {
-			switch strings.ToLower(string(node.Val)) {
+			switch strings.ToLower(node.Val) {
 			case "on":
 				return ON
 			case "off":