@vmg 2021-03-05
Very confused about this error btw:

```
Error: go/vt/vtctl/workflow/vexec/query_planner.go:269:44: cannot use ([]byte)(params.DBName) (type []byte) as type string in argument to sqlparser.NewStrLiteral
```

this file is not on my local checkout of Vitess :ghost: