@rohit-nayak-ps 2021-03-05
> ```
> Error: go/vt/vtctl/workflow/vexec/query_planner.go:269:44: cannot use ([]byte)(params.DBName) (type []byte) as type string in argument to sqlparser.NewStrLiteral
> ```
> 
> this file is not on my local checkout of Vitess

You probably need to rebase with the latest master.